# docs/openapi.yaml
openapi: 3.0.2
servers:
  - url: https://api.mainnet.hiro.so/
    description: Mainnet
  - url: https://api.nakamoto.testnet.hiro.so/
    description: Nakamoto Testnet
  - url: https://api.testnet.hiro.so/
    description: Testnet
  - url: http://localhost:3999/
    description: Local
info:
  title: Stacks Blockchain API
  version: 'STACKS_API_VERSION'
  description: |
    Welcome to the API reference overview for the [Stacks Blockchain API](https://docs.hiro.so/stacks-blockchain-api).

    [Download Postman collection](https://hirosystems.github.io/stacks-blockchain-api/collection.json)
...
paths:
...
  /extended/v1/address/{principal}/balances:
    get:
      summary: Get account balances
      description: Retrieves total account balance information for a given Address or Contract Identifier. This includes the balances of  STX Tokens, Fungible Tokens and Non-Fungible Tokens for the account.
      tags:
        - Accounts
      operationId: get_account_balance
      parameters:
        - name: principal
          in: path
          description: Stacks address or a Contract identifier
          required: true
          schema:
            type: string
            example: "SP31DA6FTSJX2WGTZ69SFY11BH51NZMB0ZW97B5P0"
        - name: unanchored
          in: query
          description: Include transaction data from unanchored (i.e. unconfirmed) microblocks
          required: false
          schema:
            type: boolean
            example: true
            default: false
        - name: until_block
          in: query
          description: returned data representing the state up until that point in time, rather than the current block.
          required: false
          schema:
            type: string
            example: 60000
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/address/get-address-balances.schema.json
              example:
                $ref: ./api/address/get-address-balances.example.json
  /extended/v1/address/{principal}/stx:
    get:
      summary: Get account STX balance
      description: Retrieves STX token balance for a given Address or Contract Identifier.
      tags:
        - Accounts
      operationId: get_account_stx_balance
      parameters:
        - name: principal
          in: path
          description: Stacks address or a Contract identifier.
          required: true
          schema:
            type: string
            example: "SP31DA6FTSJX2WGTZ69SFY11BH51NZMB0ZW97B5P0"
        - name: unanchored
          in: query
          description: Include transaction data from unanchored (i.e. unconfirmed) microblocks.
          required: false
          schema:
            type: boolean
            example: true
            default: false
        - name: until_block
          in: query
          description: returned data representing the state up until that point in time, rather than the current block. Note - Use either of the query parameters but not both at a time.
          required: false
          schema:
            type: string
            example: 60000
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/address/get-address-stx-balance.schema.json
              example:
                $ref: ./api/address/get-address-stx-balance.example.json
  /extended/v1/address/{principal}/transactions:
    get:
      summary: Get account transactions
      description: |
          Retrieves a list of all Transactions for a given Address or Contract Identifier. More information on Transaction types can be found [here](https://docs.stacks.co/understand-stacks/transactions#types).

          If you need to actively monitor new transactions for an address or contract id, we highly recommend subscribing to [WebSockets or Socket.io](https://github.com/hirosystems/stacks-blockchain-api/tree/master/client) for real-time updates.
      deprecated: true
      tags:
        - Accounts
      operationId: get_account_transactions
      parameters:
        - name: principal
          in: path
          description: Stacks address or a Contract identifier
          required: true
          schema:
            type: string
            example: "SP31DA6FTSJX2WGTZ69SFY11BH51NZMB0ZW97B5P0"
        - name: limit
          in: query
          description: max number of account transactions to fetch
          required: false
          schema:
            type: integer
            example: 42000
        - name: offset
          in: query
          description: index of first account transaction to fetch
          required: false
          schema:
            type: integer
            example: 42000
        - name: height
          in: query
          description: Filter for transactions only at this given block height
          required: false
          schema:
            type: number
            example: 42000
        - name: unanchored
          in: query
          description: Include transaction data from unanchored (i.e. unconfirmed) microblocks
          required: false
          schema:
            type: boolean
            example: true
            default: false
        - name: until_block
          in: query
          description: returned data representing the state up until that point in time, rather than the current block. Note - Use either of the query parameters but not both at a time.
          required: false
          schema:
            type: string
            example: 60000
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/address/get-address-transactions.schema.json
              example:
                $ref: ./api/address/get-address-transactions.example.json
  /extended/v1/address/{principal}/{tx_id}/with_transfers:
    get:
      summary: Get account transaction information for specific transaction
      description: Retrieves transaction details for a given Transaction Id `tx_id`, for a given account or contract Identifier.
      deprecated: true
      tags:
        - Accounts
      operationId: get_single_transaction_with_transfers
      parameters:
        - name: principal
          in: path
          description: Stacks address or a contract identifier
          required: true
          schema:
            type: string
            example: "SP3FBR2AGK5H9QBDH3EEN6DF8EK8JY7RX8QJ5SVTE"
        - name: tx_id
          in: path
          description: Transaction id
          required: true
          schema:
            type: string
            example: "0x34d79c7cfc2fe525438736733e501a4bf0308a5556e3e080d1e2c0858aad7448"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./entities/address/transaction-with-transfers.schema.json
              example:
                $ref: ./api/address/get-address-single-transaction-with-transfers.example.json
        404:
          description: Not found
          content:
            application/json:
              example:
                $ref: ./api/errors/transaction-not-found.example.json
  /extended/v1/address/{principal}/transactions_with_transfers:
    get:
      summary: Get account transactions including STX transfers for each transaction.
      description: Retrieve all transactions for an account or contract identifier including STX transfers for each transaction.
      deprecated: true
      tags:
        - Accounts
      operationId: get_account_transactions_with_transfers
      parameters:
        - name: principal
          in: path
          description: Stacks address or a Contract identifier
          required: true
          schema:
            type: string
            example: "SP31DA6FTSJX2WGTZ69SFY11BH51NZMB0ZW97B5P0"
        - name: limit
          in: query
          description: max number of account transactions to fetch
          required: false
          schema:
            type: integer
            example: 20
        - name: offset
          in: query
          description: index of first account transaction to fetch
          required: false
          schema:
            type: integer
            example: 10
        - name: height
          in: query
          description: Filter for transactions only at this given block height
          required: false
          schema:
            type: number
            example: 66119
        - name: unanchored
          in: query
          description: Include transaction data from unanchored (i.e. unconfirmed) microblocks
          required: false
          schema:
            type: boolean
            example: true
            default: false
        - name: until_block
          in: query
          description: returned data representing the state up until that point in time, rather than the current block.
          required: false
          schema:
            type: string
            example: 60000
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/address/get-address-transactions-with-transfers.schema.json
              example:
                $ref: ./api/address/get-address-transactions-with-transfers.example.json
  /extended/v1/address/{principal}/nonces:
    get:
      summary: Get the latest nonce used by an account
      description: Retrieves the latest nonce values used by an account by inspecting the mempool, microblock transactions, and anchored transactions.
      tags:
        - Accounts
      operationId: get_account_nonces
      parameters:
        - name: principal
          in: path
          description: Stacks address
          required: true
          schema:
            type: string
            example: "SP31DA6FTSJX2WGTZ69SFY11BH51NZMB0ZW97B5P0"
        - name: block_height
          in: query
          description: Optionally get the nonce at a given block height.
          required: false
          schema:
            type: number
            example: 66119
        - name: block_hash
          in: query
          description: Optionally get the nonce at a given block hash. Note - Use either of the query parameters but not both at a time.
          required: false
          schema:
            type: string
            example: "0x72d53f3cba39e149dcd42708e535bdae03d73e60d2fe853aaf61c0b392f521e9"

      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./entities/address/address-nonces.schema.json
              example:
                $ref: ./entities/address/address-nonces.example.json
  /extended/v1/address/{principal}/assets:
    get:
      summary: Get account assets
      description: Retrieves a list of all assets events associated with an account or a Contract Identifier. This includes Transfers, Mints.
      tags:
        - Accounts
      operationId: get_account_assets
      parameters:
        - name: principal
          in: path
          description: Stacks address or a Contract identifier
          required: true
          schema:
            type: string
            example: "SP31DA6FTSJX2WGTZ69SFY11BH51NZMB0ZW97B5P0"
        - name: limit
          in: query
          description: max number of account assets to fetch
          required: false
          schema:
            type: integer
            example: 20
        - name: offset
          in: query
          description: index of first account assets to fetch
          required: false
          schema:
            type: integer
            example: 42000
        - name: unanchored
          in: query
          description: Include transaction data from unanchored (i.e. unconfirmed) microblocks
          required: false
          schema:
            type: boolean
            example: true
            default: false
        - name: until_block
          in: query
          description: returned data representing the state at that point in time, rather than the current block. Note - Use either of the query parameters but not both at a time.
          required: false
          schema:
            type: string
            example: 60000
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/address/get-address-assets.schema.json
              example:
                $ref: ./api/address/get-address-assets.example.json
...